requirements.txt
pyautogui==0.9.54
Pillow>=10.1.0
google-generativeai>=0.8.0
tkinter-tooltip>=2.2.0